requirements.txt
asgiref==3.5.1
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Cython==0.29.28
dango==0.0.1
dartsclone==0.9.0
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-heroku==0.3.1
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.2.1
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
keyring==23.5.0
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.3.0
popular-times==0.1
populartimes==2.0
psycopg2==2.9.3
pycparser==2.21
pygtrie==2.4.2
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.27.1
sortedcontainers==2.1.0
sqlparse==0.4.2
SudachiDict-core==20211220
SudachiPy==0.5.4
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
zipp==3.8.0
